# Conflicts:
#	src/main/java/dao/impl/VideoDaoImpl.java
#	src/main/java/service/VideoService.java
#	src/main/java/service/impl/VideoServiceImpl.java
#	web/pages/video/historypage.jsp
#	web/pages/video/searchpage.jsp
#	web/pages/video/singlevideopage.jsp

+++ src/main/java/service/OrderService.java
@@ -17,7 +17,7 @@ public interface OrderService {
      * @param id
      * @return
      */
-    public Order queryById(Integer id);
+    public Order queryById(Long id);
 
     /**
      * 遍历订单信息
